fix: change token to cookies refactor: change routes names for login and logout
