--- api/authentication/authRoute.js
@@ -1,79 +1,0 @@
-import express from 'express';
-import bcryptjs from 'bcryptjs';
-import jsonwebtoken from 'jsonwebtoken';
-import config from 'config';
-import { check, validationResult } from 'express-validator';
-import auth from './authMiddleware.js';
-import User from '../user/User.js';
-
-const router = express.Router();
-
-// @route    GET api/auth
-// @desc     Get user by token
-// @access   Private
-router.get('/', auth, async (req, res) => {
-  try {
-    const user = await User.findById(req.user.id).select('-password');
-    res.json(user);
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send('Server Error');
-  }
-});
-
-// @route    POST api/auth
-// @desc     Authenticate user & get token
-// @access   Public
-router.post(
-  '/',
-  check('email', 'Please include a valid email').isEmail(),
-  check('password', 'Password is required').exists(),
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const { email, password } = req.body;
-
-    try {
-      const user = await User.findOne({ email });
-
-      if (!user) {
-        return res
-          .status(400)
-          .json({ errors: [{ msg: 'Invalid Credentials' }] });
-      }
-
-      const isMatch = await bcryptjs.compare(password, user.password);
-
-      if (!isMatch) {
-        return res
-          .status(400)
-          .json({ errors: [{ msg: 'Invalid Credentials' }] });
-      }
-      // If users isAdmin===true => redirect to admin panel
-      const payload = {
-        user: {
-          id: user.id,
-          isAdmin: user.isAdmin,
-        },
-      };
-
-      jsonwebtoken.sign(
-        payload,
-        config.get('jwtSecret'),
-        { expiresIn: '36000' },
-        (err, token) => {
-          if (err) throw err;
-          res.json({ token });
-        },
-      );
-    } catch (err) {
-      console.error(err.message);
-      res.status(500).send('Server error');
-    }
-  },
-);
-
-export default router;
