+++ api/user/userRoute.js
@@ -16,7 +16,7 @@ router.use('/me', userMeRoute);
 // @description Registering new user
 // @access public (anybody can register)
 router.post(
-  '/',
+  '/signup',
   check('name', 'Name is required of 5 or more characters')
     .notEmpty()
     .isString()
@@ -76,16 +76,113 @@ router.post(
         },
         (err, token) => {
           if (err) throw err;
-          res.status(200).json({ token });
+          res
+            .cookie('access_token', token, {
+              httpOnly: true,
+              sameSite: true,
+            })
+            .json({
+              isAuthenticated: true,
+              user: { id: user.id, isAdmin: user.isAdmin },
+            })
+            .status(200);
         },
       );
-      // res.status(200).send('User Created!');
     } catch (err) {
       console.error(err.message);
       res.status(500).send('Server error');
     }
   },
 );
+
+// @route    POST api/users/login
+// @desc     Authenticate user & get token
+// @access   Public
+router.post(
+  '/login',
+  check('email', 'Please include a valid email').isEmail(),
+  check('password', 'Password is required').exists(),
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+
+    const { email, password } = req.body;
+
+    try {
+      const user = await User.findOne({ email });
+
+      if (!user) {
+        return res
+          .status(400)
+          .json({ errors: [{ msg: 'Invalid Credentials' }] });
+      }
+
+      const isMatch = await bcryptjs.compare(password, user.password);
+
+      if (!isMatch) {
+        return res
+          .status(400)
+          .json({ errors: [{ msg: 'Invalid Credentials' }] });
+      }
+
+      const payload = {
+        user: {
+          id: user.id,
+          isAdmin: user.isAdmin,
+        },
+      };
+
+      jsonwebtoken.sign(
+        payload,
+        config.get('jwtSecret'),
+        { expiresIn: '36000' },
+        (err, token) => {
+          if (err) throw err;
+          res
+            .cookie('access_token', token, {
+              httpOnly: true,
+              sameSite: true,
+            })
+            .json({
+              isAuthenticated: true,
+              user: { id: user.id, isAdmin: user.isAdmin },
+            })
+            .status(200);
+        },
+      );
+    } catch (err) {
+      console.error(err.message);
+      res.status(500).send('Server error');
+    }
+  },
+);
+
+// @route GET api/users/logout
+// @description Logging out user
+// @access public (anybody can log out)
+router.get('/logout', async (req, res) => {
+  try {
+    if (!req.cookies.access_token) {
+      return res
+        .status(400)
+        .json({ errors: [{ msg: 'No cookies to clear' }] });
+    }
+    res
+      .clearCookie('access_token')
+      .json({
+        msg: 'user logged out',
+        user: { id: '', isAdmin: '' },
+        success: true,
+        isAuthenticated: false,
+      })
+      .status(200);
+  } catch (err) {
+    console.error(err.message);
+    res.status(500).send('Server Error');
+  }
+});
 
 // @route GET api/users
 // @description Getting all users
@@ -94,7 +191,7 @@ router.get('/', authMiddleware, adminMiddleware, async (req, res) => {
   try {
     const users = await User.find().select('-password');
     if (!users) res.status(404).send('Users not found');
-    res.json(users);
+    res.json({ users: users, isAuthenticated: true });
   } catch (err) {
     console.error(err.message);
     res.status(500).send('Server Error');
@@ -112,7 +209,7 @@ router.delete(
     try {
       const user = await User.findByIdAndDelete(req.body.id);
       if (!user) res.status(404).send('User not found');
-      res.status(200).json(user);
+      res.status(200).json({ user: user, isAuthenticated: true });
     } catch (err) {
       console.error(err.message);
       res.status(500).send('Server Error');
