+++ client/package.json
@@ -39,7 +39,7 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:5000",
+  "proxy": "http://127.0.0.1:5000/",
   "devDependencies": {
     "eslint": "^7.21.0",
     "eslint-config-airbnb": "^18.2.1",
